Fix coordinate type to be float
Because coordinates in the OpenGL visualization are all float, it's
important that client code maintains coordinate data as Vector<float>.

This makes a small change in the signature of the updateData and
updateCoords methods which took vector<Vector<T>>* _coords args. If
client code wants to compute coordinates in doubles, it'll have to
convert to floats before updateData/updateCoords calls are made.