Resolves Scaling issues
This commmit sorts out a problem that existed in the Scale class where
if the input range did not vary, a divide by error would occur and the
value returned by the Scale::transform would be nan, rather than the
middle of the output range.

If the inputs to ColourMap::convert are genuinely nan, then a 'nan
colour' is now generated - this can be different for each
ColourMapType, so that the 'nan colour' is always different to the
other colours in the map.

In GraphVisual some debugging is reduced.