Revert scene now takes you back to the programmer-set (or JSON set) scenetrans
I've been meaning to do this for a while. Instead scenetrans_default
being hardcoded for every morph::Visual, scenetrans_default is set
when you, the programmer, calls setSceneX() setSceneY() and
setSceneZ(). When the user presses 'a' to reset the scene, it goes
back to whatever was set by those accessors. If the user has a
/tmp/Visual.json to set up a user-preferred scene, then those will
override the programmer-provided values.