Adds 'Config overrides' to morph::Config
A morph::Config object allows the user to define a set of parameters
in a JSON file and get access to the parameters via a convenient
interface. It manages the file and provides the ability to change or
add parameters and then re-save the parameter set. This can be used to
record the parameters used for a particular simulation run.

When running batch simulations from a script, it is often useful to
scan through a parameter, in sequence. In this use case, it is
tiresome to create a new JSON file for each individual simulation run;
it would be better to use a single JSON file with all the main
parameters, and then be able to override one parameter for each run.

That's what this commit adds!

Here's what you do:

When you have created your morph::Config object, you pass argc and
argv to morph::Config::process_args(argc, argv). This then records any
overrides that you've placed into the command line of your
program. You have to use a fixed format for the overrides. So, in an
example program where you want to update the parameter 'A' in your
config, you'd run like this:

myprog any args you already use -co:A=4.5

The '-co:' is the token that Config::process_args uses to recognise
that you're overriding the value of A that is given in your existing
Config.

You can then run several programs:

myprog -co:A=4.6
myprog -co:A=4.7

and so on.

You can also use multiple instances, and specify texts and booleans,
like this:

myprog -co:A=4 -co:B=5 -co:C=text -co:D=false