Qt windows are go!
Major changes. As part of this work, I decided to modify the way that
VisualModels are set up. Now, instead of passing a shader program in
the VisualModel constructor, I set up the access to shaders via a
Visual::bindmodel(unique_ptr<VisualModel> vm) call which binds some
callbacks and sets a pointer to the parent Visual inside the
VisualModel. This is complete for GraphVisual, but lots of plumbing is
required yet...

Having done this, I have also now figured out how to get a Visual
scene rendering in a Qt window! Progress...