Go with public methods
If the function is to be unit tested, then it's simplest to keep it part of the public interface of the Grid class. 

If you want to unit test private parts of a class, you probably need to write a new class specifically for that purpose, which would provide access to the private members (using the friend keyword). But we're not going to do that. There's no great harm in making these two functions public, so I hereby reverse our previous decision.